+++ lib/src/widgets/multi_day/multi_day_body.dart
@@ -85,7 +85,7 @@ class MultiDayBody extends StatelessWidget {
                             location: context.location,
                           ),
                         ),
-                        PositionedTimeIndicator(
+                        TimeIndicatorPositioner(
                           viewController: viewController,
                           initialPage: viewController.initialPage,
                         ),